requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
dnspython==2.7.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.1.0
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.11.1
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==1.41.0
neo4j==5.28.1
neomodel==5.4.5
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.5
psutil==6.1.0
pure_eval==0.2.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rich==14.1.0
rpds-py==0.25.1
scikit-fuzzy==0.5.0
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.45.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
